requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.10.2+cu113
torchvision==0.11.3+cu113
tokenizers==0.11.6
transformers==4.16.2
Pillow==9.2.0
tqdm
numpy
git+https://github.com/rwightman/pytorch-image-models.git
wandb
datasets
jsonlines
scikit-learn

